63: * move MongoDbService to SqlService project
