+++ backend/Squirrel.Core/Squirrel.Core.BLL/Squirrel.Core.BLL.csproj
@@ -13,7 +13,6 @@
     <PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="11.0.0" />
     <PackageReference Include="Google.Apis.Auth" Version="1.61.0" />
     <PackageReference Include="DiffPlex" Version="1.7.1" />
-    <PackageReference Include="MongoDB.Driver" Version="2.21.0" />
   </ItemGroup>
 
   <ItemGroup>
