+++ backend/Squirrel.Core/Squirrel.Core.WebAPI/appsettings.json
@@ -2,10 +2,6 @@
   "ConnectionStrings": {
     "SquirrelCoreDBConnection": "Server=localhost;Database=SquirrelCoreDB;Trusted_Connection=True;"
   },
-  "MongoDatabase": {
-    "ConnectionString": "",
-    "DatabaseName": ""
-  },
   "Logging": {
     "LogLevel": {
       "Default": "Information",
