+++ backend/Squirrel.SqlService/Squirrel.SqlService.WebApi/Services/MongoService.cs
@@ -1,10 +1,10 @@
 ﻿using Microsoft.Extensions.Options;
 using MongoDB.Driver;
-using Squirrel.Core.BLL.Interfaces;
-using Squirrel.Core.Common.Models;
 using Squirrel.Core.DAL.Entities.Common;
+using Squirrel.SqlService.WebApi.Interfaces;
+using Squirrel.SqlService.WebApi.Options;
 
-namespace Squirrel.Core.BLL.Services;
+namespace Squirrel.SqlService.WebApi.Services;
 public class MongoService<T> : IMongoService<T> where T : Entity<long>
 {
     private readonly IMongoCollection<T> _mongoCollection;
