+++ backend/Squirrel.SqlService/Squirrel.SqlService.WebApi/appsettings.json
@@ -2,6 +2,10 @@
   "ConnectionStrings": {
     "SquirrelCoreDBConnection": "Server=localhost;Database=SquirrelCoreDB;Trusted_Connection=True;"
   },
+  "MongoDatabase": {
+    "ConnectionString": "",
+    "DatabaseName": ""
+  },
   "CoreWebAPIDomain": "https://localhost:44334",
   "Logging": {
     "LogLevel": {
